moves routes into public and protected routes to prevent infinite looping problem caused by protecting the route that we route to when you are not authenticated...
