+++ webapp/src/main/webapp/src/app/admin/embargo/embargo.routes.ts
@@ -1,0 +1,24 @@
+import { AuthorizationCanActivate } from '../../shared/security/authorization.can-activate';
+import { EmbargoComponent } from './embargo.component';
+import { EmbargoResolve } from './embargo.resolve';
+import { Route } from '@angular/router';
+
+export const embargoRoutes: Route[] = [
+  {
+    path: 'embargoes',
+    pathMatch: 'prefix',
+    data: {
+      breadcrumb: { translate: 'embargo.title' },
+      permissions: ['EMBARGO_WRITE']
+    },
+    canActivate: [AuthorizationCanActivate],
+    children: [
+      {
+        path: '',
+        pathMatch: 'prefix',
+        component: EmbargoComponent,
+        resolve: { embargoesByOrganizationType: EmbargoResolve }
+      }
+    ]
+  }
+];
