+++ webapp/src/main/webapp/src/app/admin/instructional-resource/instructional-resource.module.ts
@@ -34,7 +34,6 @@ import { TableModule } from 'primeng/table';
     TableModule,
     TypeaheadModule.forRoot()
   ],
-  exports: [InstructionalResourceComponent],
   providers: [
     AssessmentService,
     InstructionalResourceService,
