+++ webapp/src/main/webapp/src/app/home/home.module.ts
@@ -9,7 +9,6 @@ import { BsDropdownModule } from 'ngx-bootstrap';
 import { AdminToolsComponent } from './admin-tools.component';
 
 @NgModule({
-  declarations: [HomeComponent, AdminToolsComponent],
   imports: [
     BrowserModule,
     CommonModule,
@@ -18,6 +17,6 @@ import { AdminToolsComponent } from './admin-tools.component';
     GroupsModule,
     BsDropdownModule.forRoot()
   ],
-  exports: [HomeComponent]
+  declarations: [AdminToolsComponent, HomeComponent]
 })
 export class HomeModule {}
