+++ webapp/src/main/webapp/src/app/home/home.routes.ts
@@ -1,0 +1,10 @@
+import { HomeComponent } from './home.component';
+import { Route } from '@angular/router';
+
+export const homeRoutes: Route[] = [
+  {
+    path: '',
+    pathMatch: 'full',
+    component: HomeComponent
+  }
+];
