--- webapp/src/main/webapp/src/app/shared/security/user.resolve.ts
@@ -1,21 +1,0 @@
-import { Injectable } from '@angular/core';
-import {
-  ActivatedRouteSnapshot,
-  Resolve,
-  RouterStateSnapshot
-} from '@angular/router';
-import { Observable } from 'rxjs';
-import { UserService } from './user.service';
-import { User } from './user';
-
-@Injectable()
-export class UserResolve implements Resolve<User> {
-  constructor(private service: UserService) {}
-
-  resolve(
-    route: ActivatedRouteSnapshot,
-    state: RouterStateSnapshot
-  ): Observable<User> {
-    return this.service.getUser();
-  }
-}
