+++ webapp/src/main/webapp/src/app/user-group/user-group.module.ts
@@ -1,5 +1,4 @@
 import { NgModule } from '@angular/core';
-import { UserGroupService } from './user-group.service';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { TableModule } from 'primeng/table';
 import { BrowserModule } from '@angular/platform-browser';
@@ -7,12 +6,10 @@ import { CommonModule } from '../shared/common.module';
 import { ModalModule, PopoverModule } from 'ngx-bootstrap';
 import { UserGroupResolve } from './user-group.resolve';
 import { UserGroupComponent } from './user-group.component';
-import { UserGroupOptionsService } from './user-group-options.service';
 import { StudentSearchModule } from '../student/search/student-search.module';
 import { UserGroupFormComponent } from './user-group-form.component';
 import { UserGroupsComponent } from './user-groups.component';
 import { UserGroupTableComponent } from './user-group-table.component';
-import { ExamModule } from '../exam/exam.module';
 
 @NgModule({
   declarations: [
@@ -31,7 +28,7 @@ import { ExamModule } from '../exam/exam.module';
     StudentSearchModule,
     TableModule
   ],
-  exports: [UserGroupComponent, UserGroupsComponent],
-  providers: [UserGroupService, UserGroupOptionsService, UserGroupResolve]
+  providers: [UserGroupResolve],
+  exports: [UserGroupsComponent]
 })
 export class UserGroupModule {}
